Issue#249: documentation, code layout changed
