+++ config/src/test/resources/beanconfig/beanconfig01.conf
@@ -55,8 +55,7 @@
     },
     "durations" : {
         "second" : 1s,
-        "secondsList" : [1s,2seconds,3 s, 4000],
-        "secondAsNumber" : 1000,
+        "secondAsNumber" : 1000000000,
         "halfSecond" : 0.5s
     },
