+++ config/src/test/scala/com/typesafe/config/ConfigBeanFactoryTest.scala
@@ -1,14 +1,12 @@
 /**
  * Copyright (C) 2013 Typesafe Inc. <http://typesafe.com>
  */
-package com.typesafe.config.impl
+package com.typesafe.config
 
 import java.io.{InputStream, InputStreamReader}
 import java.util.concurrent.TimeUnit
 
 import beanconfig._
-import com.typesafe.config.ConfigException.Generic
-import com.typesafe.config._
 import org.junit.Assert._
 import org.junit._
 
@@ -73,12 +71,18 @@ class ConfigBeanFactoryTest {
   def testCreateDuration() {
     val beanConfig: DurationsConfig = ConfigBeanFactory.create(loadConfig().getConfig("durations"), classOf[DurationsConfig])
     assertNotNull(beanConfig)
+    assertEquals(beanConfig.getHalfSecond,TimeUnit.MILLISECONDS.toNanos(500))
+    assertEquals(beanConfig.getSecond,TimeUnit.SECONDS.toNanos(1))
+    assertEquals(beanConfig.getSecondAsNumber,TimeUnit.SECONDS.toNanos(1))
   }
 
   @Test
   def testCreateBytes() {
     val beanConfig: BytesConfig = ConfigBeanFactory.create(loadConfig().getConfig("bytes"), classOf[BytesConfig])
     assertNotNull(beanConfig)
+    assertEquals(beanConfig.getKibibyte,1024L)
+    assertEquals(beanConfig.getKilobyte,1000L)
+    assertEquals(beanConfig.getThousandBytes,1000L)
   }
 
   private def loadConfig(): Config = {
